some basic fleshing out...
