+++ pyspinw/experiment.py
@@ -1,5 +1,19 @@
 """Different kinds of experimental setups"""
 
-#pylint: disable=W0611
-from pyspinw._base import Experiment
-#pylint: enable=W0611
+from pyspinw.sample import Sample
+from pyspinw._base import Data
+
+class Experiment:
+    """The setup of a neutron experiment."""
+
+    def __init__(self, sample: Sample, instrument: Instrument | None = None):
+        self.sample = sample
+        self.instrument = instrument
+
+    def calculate(self):
+        """Calculate energies for the experiment."""
+        raise NotImplementedError
+
+    def fit(self, data: Data):
+        """Fit the experimental model to a dataset."""
+        raise NotImplementedError
