+++ pyspinw/sample.py
@@ -1,8 +1,20 @@
 """Different Kinds of samples"""
 
-from pyspinw._base import Sample
+from abc import ABC, abstractmethod
+
+from pyspinw.hamiltonian import Hamiltonian
 
 # pylint: disable=R0903
+
+class Sample(ABC):
+    """Representation of the macrostructure of a sample used in an experiment (Twin, Powder etc)"""
+
+    def __init__(self, hamiltonian: Hamiltonian):
+        self.hamiltonian = hamiltonian
+
+    @abstractmethod
+    def generate_q(self, n_q: int):
+        """Generate a sample of q-vectors for a given magnetic field and resolution."""
 
 
 class SingleCrystal(Sample):
